+++ chess/rook.py
@@ -2,24 +2,13 @@ from chess.pieces import Piece       #TORRE CLASE
 
 class Rook(Piece):
 
-    white_str = "♜"
-    black_str = "♖"
+    def symbol(self):
+        return 'R' if self.get_color() == "WHITE" else 'r'
 
-    def valid_positions(
-        self,
-        from_row,
-        from_col,
-        to_row,
-        to_col,
-    ):
+# Movimientos ortogonales    
 
-        possible_positions = (
-            self.possible_positions_vd(from_row, from_col) +  # verticales hacia abajo
-            self.possible_positions_va(from_row, from_col) +  # verticales hacia arriba
-            self.possible_positions_hr(from_row, from_col) +  # horizontales hacia la derecha
-            self.possible_positions_hl(from_row, from_col)    # horizontales hacia la izquierda
-        )
+    def is_valid_piece_move(self, from_row, from_col, to_row, to_col):
+        # Movimientos ortogonales: arriba, abajo, izquierda, derecha
+        directions = [(1, 0), (-1, 0), (0, -1), (0, 1)]
+        possible_positions = self.calculate_possible_moves(from_row, from_col, directions)
         return (to_row, to_col) in possible_positions
-
-
-# agregar movimientos ortogonales (hecho)